backend/requirements.txt
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.6
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.15.1
Flask==3.0.3
Flask-Cors==4.0.0
h11==0.14.0
harness-featureflags==1.6.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mmh3==4.1.0
PyJWT==2.8.0
redis==5.0.4
requests==2.31.0
sniffio==1.3.1
sqlparse==0.5.0
tenacity==8.2.2
urllib3==2.2.1
Werkzeug==3.0.2
